backup_arquivos_antigos/20251106_202754/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
lxml==4.9.3
pandas==2.1.3
openpyxl==3.1.2

